[equals][overlaps] Divide the code into interface and implementation part.
